requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
aiofiles
